Chisel Outdoors Reboot
Initial commit of chiseloutdoors.com. This is a reboot of the site
utilizing the techniques and methods learned from the Udacity Fullstack
Nanodegree program.

This is a direct copy of the basic catalog project with Facebook and
Google OAuth login. Testing via Vagrant. Current database is sqlLite.
Login and user permission tested and working.

Next step is to incorporate the search page functionality from the
Knockout.js project with animated markers and menu screens. After which
will be to query those marker locations from prior database entries.