Destination
Structure - Selected parks, trails, and content now fall into
‘destination’ binding which is used to display the destination name and
in the Controller for the relevant position for searches and weather
queries.

* Need a solution for when a trail is selected followed by a ‘places’
search since the destination title currently stays the same but all
trails are cleared and only the relevant park is left on the map. Will
need to clear all trails other than the currently selected and the
relevant park, if any.